[bluetooth.generic] Preserve decimals when writing integer fields with a DecimalExponent
A GATT integer field can carry a DecimalExponent, which makes it represent
a fractional real-world value: the Temperature characteristic (0x2A6E), for
example, is a sint16 in units of 0.01 degrees C.

When writing such a field, the state was converted with DecimalType#longValue(),
which truncates the fraction before the field multiplier is applied. Writing
21.5 degrees therefore serialized to a raw 2100 (21.00 degrees) instead of 2150,
silently losing everything after the decimal point.

Use DecimalType#doubleValue() instead. GattRequest#setField(String, Double)
divides by the field multiplier before encoding, so the fractional part
survives into the raw integer value that is written to the device.

+++ bundles/org.openhab.binding.bluetooth.generic/src/main/java/org/openhab/binding/bluetooth/generic/internal/BluetoothChannelUtils.java
@@ -158,7 +158,12 @@ public class BluetoothChannelUtils {
                             request.getCharacteristicUUID(), fieldName, state);
                     return;
                 }
-                request.setField(fieldName, decimalType.longValue());
+                // Use doubleValue(), not longValue(): when the integer field carries a
+                // DecimalExponent the real-world value can be fractional (e.g. 21.5 for a
+                // sint16 in 0.01 degC units). setField(Double) divides by the field multiplier
+                // (10^exponent) before encoding, yielding the correct raw integer, whereas
+                // longValue() truncates the fraction first and corrupts the written value.
+                request.setField(fieldName, decimalType.doubleValue());
                 return;
             }
             case FLOAT_IEE754:
